Group appdevelopment tests
