+++ backend/tests/Designer.Tests/Controllers/AppDevelopmentController/AppDevelopmentControllerTestsBase.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using Altinn.Studio.Designer.Configuration;
+using Altinn.Studio.Designer.Services.Interfaces;
+using Designer.Tests.Controllers.ApiTests;
+using Designer.Tests.Mocks;
+using Designer.Tests.Utils;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Designer.Tests.Controllers.AppDevelopmentController
+{
+    public class AppDevelopmentControllerTestsBase<TControllerTestType> : ApiTestsBase<Altinn.Studio.Designer.Controllers.AppDevelopmentController, TControllerTestType>, IDisposable
+    where TControllerTestType : class
+    {
+        protected static string VersionPrefix(string org, string repository) => $"/designer/api/{org}/{repository}/app-development";
+        protected string CreatedFolderPath { get; set; }
+
+        public void Dispose()
+        {
+            if (!string.IsNullOrWhiteSpace(CreatedFolderPath))
+            {
+                TestDataHelper.DeleteDirectory(CreatedFolderPath);
+            }
+        }
+
+        public AppDevelopmentControllerTestsBase(WebApplicationFactory<Altinn.Studio.Designer.Controllers.AppDevelopmentController> factory) : base(factory)
+        {
+        }
+
+        protected override void ConfigureTestServices(IServiceCollection services)
+        {
+            services.Configure<ServiceRepositorySettings>(c =>
+                c.RepositoryLocation = TestRepositoriesLocation);
+            services.AddSingleton<IGitea, IGiteaMock>();
+        }
+
+    }
+}
